requirements.txt
websocket-client==1.8.0
dotenv==0.9.9